feature: Se agrega logica de negocio pra que el servicio este siempre asociado a un tecnico feature: Se configura GlobalExceptionHandler y se agrega un error response para manejar business exception

+++ domain/model/src/main/java/co/com/bancolombia/model/service/Service.java
@@ -16,5 +16,5 @@ public class Service {
     private String description;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
-    private Long technicianId;
+    private Integer technicianId;
 }
